+++ snooty/types.py
@@ -331,6 +331,7 @@ class ProjectConfig:
     substitutions: Dict[str, str] = field(default_factory=dict)
     # substitution_nodes contains a parsed representation of the substitutions member, and is populated on Project initialization.
     substitution_nodes: Dict[str, SerializableType] = field(default_factory=dict)
+    toc_landing_pages: List[str] = field(default_factory=list)
 
     @property
     def source_path(self) -> Path: